+++ lib/titlebar_buttons.dart
@@ -1,4 +1,4 @@
-library window_decorations;
+library titlebar_buttons;
 
 export 'src/decorated_button.dart';
 export 'src/theme_type.dart';